+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   - Make output better structured
   - Add -PassThru to Invoke-Pester [GH-102], [GH-84] and [GH-46]
   - Makes New-Fixture -Path option more resilient [GH-114]
+  - Make the New-Fixture input accept any path and output objects
   
 ## 2.0.3 (Apr 16, 2013)